offsetMinutes

+++ app/src/main/java/com/forrestguice/suntimeswidget/map/WorldMapDialog.java
@@ -19,7 +19,6 @@
 package com.forrestguice.suntimeswidget.map;
 
 import android.annotation.SuppressLint;
-import android.app.Activity;
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -33,10 +32,8 @@ import android.support.annotation.Nullable;
 import android.support.design.widget.BottomSheetBehavior;
 import android.support.design.widget.BottomSheetDialog;
 import android.support.design.widget.BottomSheetDialogFragment;
-import android.support.v4.app.DialogFragment;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.widget.ImageViewCompat;
-import android.support.v7.app.AlertDialog;
 
 import android.support.v7.widget.PopupMenu;
 import android.util.Log;
@@ -50,7 +47,6 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.FrameLayout;
 import android.widget.ImageButton;
-import android.widget.MediaController;
 import android.widget.RadioButton;
 import android.widget.SeekBar;
 import android.widget.Spinner;
@@ -598,7 +594,7 @@ public class WorldMapDialog extends BottomSheetDialogFragment
     private WorldMapTask.WorldMapTaskListener onWorldMapUpdate = new WorldMapTask.WorldMapTaskListener()
     {
         @Override
-        public void onFrame(Bitmap result, int offsetMinutes)
+        public void onFrame(Bitmap result, long offsetMinutes)
         {
             expandSheet(getDialog());
             if (seekbar != null)
